Moved PistonBaseBlockMixing logic from fabric/1611 to fabric/common
