--- fabric/mc1161/src/main/resources/grimac-mc1161.mixins.json
@@ -1,12 +1,0 @@
-{
-    "required": true,
-    "minVersion": "0.8",
-    "package": "ac.grim.grimac.platform.fabric.mc1161.mixins",
-    "compatibilityLevel": "JAVA_17",
-    "mixins": [
-        "PistonBaseBlockMixin"
-    ],
-    "injectors": {
-        "defaultRequire": 1
-    }
-}
