+++ fabric/src/main/resources/grimac.mixins.json
@@ -6,7 +6,8 @@
     "mixins": [
         "LevelChunkMixin",
         "LevelMixin",
-        "ServerPlayerMixin"
+        "ServerPlayerMixin",
+        "PistonBaseBlockMixin"
     ],
     "injectors": {
         "defaultRequire": 1
